refactor: PostCommandFacade로 리팩터링
