+++ src/main/java/com/soolsul/soolsulserver/post/facade/PostFacadeGateway.java
@@ -1,7 +1,6 @@
-package com.soolsul.soolsulserver.post.business;
+package com.soolsul.soolsulserver.post.facade;
 
 import com.soolsul.soolsulserver.location.request.LocationSquareRangeRequest;
-import com.soolsul.soolsulserver.post.facade.PostQueryFacade;
 import com.soolsul.soolsulserver.post.presentation.dto.PostCreateRequest;
 import com.soolsul.soolsulserver.post.presentation.dto.PostDetailResponse;
 import com.soolsul.soolsulserver.post.presentation.dto.PostListResponse;
@@ -11,14 +10,14 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class PostServiceGateway implements CommandService<PostCreateRequest>, QueryService {
+public class PostFacadeGateway implements CommandFacadeSpec<PostCreateRequest>, QueryFacadeSpec {
 
-    private final PostCommandService postCommandService;
+    private final PostCommandFacade postCommandFacade;
     private final PostQueryFacade postQueryFacade;
 
     @Override
     public void create(String userId, PostCreateRequest request) {
-        postCommandService.create(userId, request);
+        postCommandFacade.create(userId, request);
     }
 
     @Override
