+++ src/main/java/com/soolsul/soolsulserver/post/facade/QueryFacadeSpec.java
@@ -1,11 +1,11 @@
-package com.soolsul.soolsulserver.post.business;
+package com.soolsul.soolsulserver.post.facade;
 
 import com.soolsul.soolsulserver.location.request.LocationSquareRangeRequest;
 import com.soolsul.soolsulserver.post.presentation.dto.PostDetailResponse;
 import com.soolsul.soolsulserver.post.presentation.dto.PostListResponse;
 import org.springframework.data.domain.Pageable;
 
-public interface QueryService {
+public interface QueryFacadeSpec {
     PostDetailResponse find(String userId, String id);
 
     PostListResponse findAll(String userId, LocationSquareRangeRequest locationSquareRangeRequest, Pageable pageable);
