+++ src/test/java/com/soolsul/soolsulserver/unit/post/business/PostQueryFacadeSpecTest.java
@@ -27,7 +27,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
-public class PostQueryServiceTest {
+public class PostQueryFacadeSpecTest {
 
     @InjectMocks
     private PostQueryService postQueryService;
